microservices/api-interface/requirements.txt
flask ~= 2.2.3
grpcio-tools ~= 1.30
Jinja2 ~= 3.1.2